Add EncodingUtil
