+++ src/main/java/net/dv8tion/jda/internal/requests/restaction/AuditableRestActionImpl.java
@@ -21,7 +21,7 @@ import net.dv8tion.jda.api.audit.ThreadLocalReason;
 import net.dv8tion.jda.api.requests.Request;
 import net.dv8tion.jda.api.requests.Response;
 import net.dv8tion.jda.api.requests.restaction.AuditableRestAction;
-import net.dv8tion.jda.api.utils.MiscUtil;
+import net.dv8tion.jda.api.utils.EncodingUtil;
 import net.dv8tion.jda.internal.requests.RestActionImpl;
 import net.dv8tion.jda.internal.requests.Route;
 import okhttp3.RequestBody;
@@ -110,7 +110,7 @@ public class AuditableRestActionImpl<T> extends RestActionImpl<T> implements Aud
 
     private String uriEncode(String input)
     {
-        String formEncode = MiscUtil.encodeUTF8(input);
+        String formEncode = EncodingUtil.encodeUTF8(input);
         return formEncode.replace('+', ' ');
     }
 }
